pyproject.toml
[tool.poetry]
name = "shadems"
version = "0.5.2"
description = "Rapid Measurement Set plotting with dask-ms and datashader"
authors = ["Ian Heywood & RATT <ian.heywood@physics.ox.ac.uk>"]
license = "GPL-2.0-only"
homepage = "https://github.com/ratt-ru/shadeMS"
repository = "https://github.com/ratt-ru/shadeMS"
readme = "README.md"
keywords = ["Astronomy", "Visualisation", "Packaging"]
classifiers = [
        "Development Status :: 4 - Beta",
        "Intended Audience :: Science/Research",
        "License :: OSI Approved :: GNU General Public License v2 (GPLv2)",
        "Operating System :: POSIX :: Linux",
        "Programming Language :: Python",
        "Topic :: Scientific/Engineering :: Astronomy"
]
packages = [
    { include = "shade_ms" },
]

[tool.poetry.scripts]
shadems = {reference = "bin/shadems", type = "file"}

[tool.poetry.dependencies]
python = "^3.8"
datashader = "^0.15.0"
dask-ms = { version = "^0.2.16", extras = ["xarray"] }
holoviews = "^1.14.9"
matplotlib = { version = "^3.6.0" }
cmasher = "^1.6.3"
future-fstrings = "^1.2.0"
requests = "^2.27.1"
pytest = { version = "^7.2.2", optional=true }
pytest-flake8 = { version = "^1.1.1", optional=true }

[tool.poetry.extras]
testing = ["pytest", "pytest-flake8"]

[build-system]
requires = ["setuptools", "poetry-core"]
build-backend = "poetry.core.masonry.api"